[refactor] category member 추가 및 모델 이동
