+++ app/src/main/java/com/mashup/zuzu/data/response/GetDrinksResponse.kt
@@ -1,6 +1,7 @@
 package com.mashup.zuzu.data.response
 
 import com.google.gson.annotations.SerializedName
+import com.mashup.zuzu.data.response.model.Category
 
 /**
  * @Created by 김현국 2022/08/08
@@ -32,6 +33,9 @@ data class GetDrinksResponse(
     val description: String,
 
     @SerializedName("image_url")
-    val image_url: String
+    val image_url: String,
+
+    @SerializedName("category")
+    val category: Category
 
 )
