+++ app/src/main/java/com/mashup/zuzu/data/response/model/Category.kt
@@ -1,0 +1,12 @@
+package com.mashup.zuzu.data.response.model
+
+import com.google.gson.annotations.SerializedName
+
+/**
+ * @Created by 김현국 2022/08/11
+ */
+data class Category(
+
+    @SerializedName("name")
+    val name: String
+)
