+++ app/src/main/java/com/mashup/zuzu/data/response/model/Profile.kt
@@ -1,0 +1,16 @@
+package com.mashup.zuzu.data.response.model
+
+import com.google.gson.annotations.SerializedName
+
+/**
+ * @Created by 김현국 2022/08/11
+ */
+
+data class Profile(
+
+    @SerializedName("id")
+    val id: Long,
+
+    @SerializedName("image_url")
+    val image_url: String
+)
